refactored calling the handler
