+++ src/yet_another_retry/utils/__init__.py
@@ -1,5 +1,5 @@
-from .get_func_meta import get_func_meta
-from .filter_retry_config import filter_retry_config
 from .retry_config import RetryConfig
+from .get_func_meta import get_func_meta
+from .call_handler import call_handler
 
-__all__ = ["get_func_meta", "filter_retry_config", "RetryConfig"]
+__all__ = ["RetryConfig", "get_func_meta", "call_handler"]
